refactor: Move validation tests
Create new file 'test_exceptions.py' to contain tests for exceptions.
Move tests that verify error handling logic and raised exceptions to new file.
